Reorganize and expand comments for rawPrivateKeyOf
